Fixes rounding errors, adds Province string
In case of checksum mismatches due to rounding errors the following strategy is applied:
If the checksum is higher then the number of children the value with the smallest difference to
the rounding boundary (x,5) is rounded to its ceiling value, e.g. in 5,321, 5,419 and 5,446 the value 5,446 isrounded up to 6,000
If the checksum is higher then the number of children the value with the smallest difference to
the rounding boundary (x,5) is rounded to its floor value, e.g. in 1,575, 3,780 and 5,607 the value 1,575 is rounded down to 1,000

There should be a general rounding strategy (whatever you pick) put in place for the project. You might want to put this into the README.md.